整理结构
